# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 수영장 관리 플랫폼 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pools/apply:
    post:
      operationId: applyPool
      tags: [pools]
      summary: 수영장 가입 신청
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyPoolRequest"
      responses:
        "201":
          description: 신청 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
  /pools/my:
    get:
      operationId: getMyPool
      tags: [pools]
      summary: 내 수영장 조회
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
...
  /notices:
    get:
      operationId: getNotices
      tags: [notices]
      summary: 공지사항 목록
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"
    post:
      operationId: createNotice
      tags: [notices]
      summary: 공지사항 작성
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeRequest"
      responses:
        "201":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pool:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        owner_name:
          type: string
        owner_email:
          type: string
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        rejection_reason:
          type: string
          nullable: true
        subscription_status:
          type: string
          enum: [trial, active, expired, suspended, cancelled]
        subscription_start_at:
          type: string
          nullable: true
        subscription_end_at:
          type: string
          nullable: true
        member_count:
          type: integer
          nullable: true
        created_at:
          type: string
      required:
        - id
        - name
        - address
        - phone
        - owner_name
        - owner_email
        - approval_status
        - subscription_status
        - created_at
    ApplyPoolRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        owner_name:
          type: string
      required:
        - name
        - address
        - phone
        - owner_name
...
    Notice:
      type: object
      properties:
        id:
          type: string
        swimming_pool_id:
          type: string
        title:
          type: string
        content:
          type: string
        author_name:
          type: string
        is_pinned:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - swimming_pool_id
        - title
        - content
        - author_name
        - is_pinned
        - created_at
    CreateNoticeRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        is_pinned:
          type: boolean
      required:
        - title
        - content